feat(key-manager): add default implementation to legacy sign methods that leverage the unified signer
